Update deps, start linux build

+++ src/mingwX64Main/kotlin/knxlauncher/putils.kt
@@ -9,10 +9,10 @@ typealias WSTR = CPointer<UShortVar>
 
 fun WSTR.toKString(): String = memScoped {
     // Figure out how much memory we need after UTF-8 conversion.
-    val sz = WideCharToMultiByte(CP_UTF8, WC_ERR_INVALID_CHARS, this@toKString, -1, null, 0, null, null)
+    val sz = WideCharToMultiByte(CP_UTF8.toUInt(), WC_ERR_INVALID_CHARS.toUInt(), this@toKString, -1, null, 0, null, null)
     // Now convert to UTF-8 and from there, a String.
     val utf8 = allocArray<ByteVar>(sz)
-    val r = WideCharToMultiByte(CP_UTF8, WC_ERR_INVALID_CHARS, this@toKString, -1, utf8, sz, null, null)
+    val r = WideCharToMultiByte(CP_UTF8.toUInt(), WC_ERR_INVALID_CHARS.toUInt(), this@toKString, -1, utf8, sz, null, null)
     if (r == 0) {
         throw RuntimeException("Could not convert to UTF-8")
     }
